+++ hil-test/tests/spi_full_duplex.rs
@@ -6,15 +6,25 @@
 #![no_main]
 
 use embedded_hal::spi::SpiBus;
+#[cfg(pcnt)]
+use esp_hal::{
+    gpio::interconnect::InputSignal,
+    pcnt::{channel::EdgeMode, unit::Unit, Pcnt},
+};
 use esp_hal::{
     gpio::Io,
+    peripherals::SPI2,
     prelude::*,
     spi::{master::Spi, FullDuplexMode, SpiMode},
 };
 use hil_test as _;
 
 struct Context {
-    spi: Spi<'static, esp_hal::peripherals::SPI2, FullDuplexMode>,
+    spi: Spi<'static, SPI2, FullDuplexMode>,
+    #[cfg(pcnt)]
+    pcnt_source: InputSignal,
+    #[cfg(pcnt)]
+    pcnt_unit: Unit<'static, 0>,
 }
 
 #[cfg(test)]
@@ -34,12 +44,23 @@ mod tests {
         let (_, mosi) = hil_test::common_test_pins!(io);
 
         let mosi_loopback = mosi.peripheral_input();
+        #[cfg(pcnt)]
+        let mosi_loopback_pcnt = mosi.peripheral_input();
         let spi = Spi::new(peripherals.SPI2, 100.kHz(), SpiMode::Mode0)
             .with_sck(sclk)
             .with_mosi(mosi)
             .with_miso(mosi_loopback);
 
-        Context { spi }
+        #[cfg(pcnt)]
+        let pcnt = Pcnt::new(peripherals.PCNT);
+
+        Context {
+            spi,
+            #[cfg(pcnt)]
+            pcnt_source: mosi_loopback_pcnt,
+            #[cfg(pcnt)]
+            pcnt_unit: pcnt.unit0,
+        }
     }
 
     #[test]
@@ -63,6 +84,44 @@ mod tests {
             .expect("Asymmetric transfer failed");
         assert_eq!(write[0], read[0]);
         assert_eq!(read[2], 0x00u8);
+    }
+
+    #[test]
+    #[timeout(3)]
+    #[cfg(pcnt)]
+    fn test_asymmetric_write(mut ctx: Context) {
+        let write = [0xde, 0xad, 0xbe, 0xef];
+
+        let unit = ctx.pcnt_unit;
+
+        unit.channel0.set_edge_signal(ctx.pcnt_source);
+        unit.channel0
+            .set_input_mode(EdgeMode::Hold, EdgeMode::Increment);
+
+        SpiBus::write(&mut ctx.spi, &write[..]).expect("Asymmetric write failed");
+        // Flush because we're not reading, so the write may happen in the background
+        ctx.spi.flush().expect("Flush failed");
+
+        assert_eq!(unit.get_value(), 9);
+    }
+
+    #[test]
+    #[timeout(3)]
+    #[cfg(pcnt)]
+    fn test_asymmetric_write_transfer(mut ctx: Context) {
+        let write = [0xde, 0xad, 0xbe, 0xef];
+
+        let unit = ctx.pcnt_unit;
+
+        unit.channel0.set_edge_signal(ctx.pcnt_source);
+        unit.channel0
+            .set_input_mode(EdgeMode::Hold, EdgeMode::Increment);
+
+        SpiBus::transfer(&mut ctx.spi, &mut [], &write[..]).expect("Asymmetric transfer failed");
+        // Flush because we're not reading, so the write may happen in the background
+        ctx.spi.flush().expect("Flush failed");
+
+        assert_eq!(unit.get_value(), 9);
     }
 
     #[test]